Fix type of rejoin cooldown variable
Changed the type of the 'cooldown' variable from long to int to match the return type of settings.getInt(REJOIN_COOLDOWN).

+++ src/main/java/net/sacredlabyrinth/phaed/simpleclans/managers/ClanManager.java
@@ -972,7 +972,7 @@ public final class ClanManager {
             return 0L;
         }
 
-        final long cooldown = settings.getInt(REJOIN_COOLDOWN);
+        final int cooldown = settings.getInt(REJOIN_COOLDOWN);
         if (cooldown <= 0) {
             return 0L;
         }
